feat: add new theme oloodi-hugo-clean
This commit introduces a new theme, oloodi-hugo-clean, to the project.  The new theme provides a clean, modern design with dark mode support and search functionality.  It replaces the previous 'bwnyasse-theme'.  A LICENSE and README file have been added for the new theme.  The necessary files for the theme have been added to the project.